requirements.txt
discord.py[voice]>=2.4.0
discord-ext-voice-recv>=0.4.2a179
numpy>=1.26.0
python-dotenv>=1.0.1
aiohttp>=3.9.5
wyoming>=1.5.4
